codes/requirements.txt
#Requirements - ScipyStack

numpy>=1.6
matplotlib>=1.1
ipython>=0.13
pandas>=0.8
sympy>=0.7
nose>=1.1
scipy>=0.1
pyside>=1.2


#Requirements - Sound Reading

pysoundfile>=0.8
